PP-6267 Move queue level classes to domain specific packages
- Moved `queue` level classes to domain specific packages
