+++ src/main/java/uk/gov/pay/connector/events/StateTransitionQueueMetricEmitter.java
@@ -3,7 +3,7 @@ package uk.gov.pay.connector.events;
 import com.codahale.metrics.Gauge;
 import com.codahale.metrics.MetricRegistry;
 import io.dropwizard.setup.Environment;
-import uk.gov.pay.connector.queue.StateTransitionQueue;
+import uk.gov.pay.connector.queue.statetransition.StateTransitionQueue;
 
 import javax.inject.Inject;
