+++ src/main/java/uk/gov/pay/connector/queue/statetransition/RefundStateTransition.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.queue;
+package uk.gov.pay.connector.queue.statetransition;
 
 import uk.gov.pay.connector.refund.model.domain.RefundStatus;
