+++ src/main/java/uk/gov/pay/connector/queue/statetransition/StateTransition.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.queue;
+package uk.gov.pay.connector.queue.statetransition;
 
 import java.util.concurrent.Delayed;
 import java.util.concurrent.TimeUnit;
