+++ src/test/java/uk/gov/pay/connector/events/model/refund/EventFactoryTest.java
@@ -39,9 +39,9 @@ import uk.gov.pay.connector.gateway.sandbox.SandboxPaymentProvider;
 import uk.gov.pay.connector.pact.ChargeEventEntityFixture;
 import uk.gov.pay.connector.pact.RefundHistoryEntityFixture;
 import uk.gov.pay.connector.paritycheck.LedgerTransaction;
-import uk.gov.pay.connector.queue.PaymentStateTransition;
-import uk.gov.pay.connector.queue.RefundStateTransition;
-import uk.gov.pay.connector.queue.StateTransition;
+import uk.gov.pay.connector.queue.statetransition.PaymentStateTransition;
+import uk.gov.pay.connector.queue.statetransition.RefundStateTransition;
+import uk.gov.pay.connector.queue.statetransition.StateTransition;
 import uk.gov.pay.connector.refund.dao.RefundDao;
 import uk.gov.pay.connector.refund.model.domain.RefundHistory;
 import uk.gov.pay.connector.refund.model.domain.RefundStatus;
