+++ src/test/java/uk/gov/pay/connector/it/events/EmittedEventResourceIT.java
@@ -12,7 +12,7 @@ import uk.gov.pay.connector.it.base.ChargingITestBase;
 import uk.gov.pay.connector.junit.ConfigOverride;
 import uk.gov.pay.connector.junit.DropwizardConfig;
 import uk.gov.pay.connector.junit.DropwizardJUnitRunner;
-import uk.gov.pay.connector.queue.StateTransitionQueue;
+import uk.gov.pay.connector.queue.statetransition.StateTransitionQueue;
 
 import java.sql.Timestamp;
 import java.time.Instant;
